Make VoteFactory choose a Keg in the Poll.
There was a validation error if you tried to instantiate
a VoteFactory without arguments because it used a Keg
that wasn't in the Poll it created.

Some tests worked around this by creating their own factory
instances for vote fields. This commit refactors those out,
partly to clean up after an old bug and partly so future
changes are less likely to require combing through tests for
factory instantiations that override default behavior.